requirements.txt
Flask==1.1.2
mysql-connector-python-rf==2.2.2